Fixed font rendering for Python 3
Fixed some math thing which prevented the font from being rendered properly in Python 3 (worked fine in Python 2 though).

+++ matabicho.py
@@ -71,7 +71,7 @@ def Texto( x, y, texto ):
     for char in texto:
         letra = ord(char)
         if not letra==' ':
-            screen.blit( font, [x,y], [32*mod(letra,16),32*(letra/16),28,32] )
+            screen.blit( font, [x,y], [32*mod(letra,16),32*int(letra/16),28,32] )
         x=x+18
 
 #
